add fourier lamino module, reorganize processing
