+++ src/cuda/CMakeLists.txt
@@ -11,8 +11,8 @@ include(${SWIG_USE_FILE})
 find_package(Python COMPONENTS Development REQUIRED)
 include_directories(${PYTHON_INCLUDE_PATH})
 # TODO: Is NumPy required?
-#find_package(NumPy REQUIRED)
-#include_directories(${NumPy_INCLUDE_DIRS})
+# find_package(NumPy REQUIRED)
+# include_directories(${NumPy_INCLUDE_DIRS})
 # Must tell CMake that SWIG interface files are CXX files
 
 set_source_files_properties(cfunc_fourierrec.i PROPERTIES CPLUSPLUS ON)
@@ -75,3 +75,27 @@ target_compile_definitions(cfunc_filterfp16 PRIVATE HALF)
 install(FILES "${CMAKE_CURRENT_BINARY_DIR}/cfunc_filterfp16.py" DESTINATION tomocupy)
 install(TARGETS cfunc_filterfp16 LIBRARY DESTINATION tomocupy)
 
+
+# Must tell CMake that SWIG interface files are CXX files
+set_source_files_properties(cfunc_usfft1d.i PROPERTIES CPLUSPLUS ON)
+swig_add_library(cfunc_usfft1d LANGUAGE python SOURCES cfunc_usfft1d.i cfunc_usfft1d.cu)
+swig_link_libraries(cfunc_usfft1d cudart cufft Python::Module)
+set_target_properties(cfunc_usfft1d PROPERTIES CUDA_ARCHITECTURES "60;72;75;80")
+install(FILES "${CMAKE_CURRENT_BINARY_DIR}/cfunc_usfft1d.py" DESTINATION tomocupy)
+install(TARGETS cfunc_usfft1d LIBRARY DESTINATION tomocupy)
+
+# Must tell CMake that SWIG interface files are CXX files
+set_source_files_properties(cfunc_usfft2d.i PROPERTIES CPLUSPLUS ON)
+swig_add_library(cfunc_usfft2d LANGUAGE python SOURCES cfunc_usfft2d.i cfunc_usfft2d.cu)
+swig_link_libraries(cfunc_usfft2d cudart cufft Python::Module)
+set_target_properties(cfunc_usfft2d PROPERTIES CUDA_ARCHITECTURES "60;72;75;80")
+install(FILES "${CMAKE_CURRENT_BINARY_DIR}/cfunc_usfft2d.py" DESTINATION tomocupy)
+install(TARGETS cfunc_usfft2d LIBRARY DESTINATION tomocupy)
+
+# Must tell CMake that SWIG interface files are CXX files
+set_source_files_properties(cfunc_fft2d.i PROPERTIES CPLUSPLUS ON)
+swig_add_library(cfunc_fft2d LANGUAGE python SOURCES cfunc_fft2d.i cfunc_fft2d.cu)
+swig_link_libraries(cfunc_fft2d cudart cufft Python::Module)
+set_target_properties(cfunc_fft2d PROPERTIES CUDA_ARCHITECTURES "60;72;75;80")
+install(FILES "${CMAKE_CURRENT_BINARY_DIR}/cfunc_fft2d.py" DESTINATION tomocupy)
+install(TARGETS cfunc_fft2d LIBRARY DESTINATION tomocupy)
