+++ src/tomocupy/backproj_functions.py
@@ -43,15 +43,12 @@ from tomocupy import lprec
 from tomocupy import fbp_filter
 from tomocupy import linerec
 from tomocupy import utils
-from tomocupy import retrieve_phase, remove_stripe, adjust_projections
 import cupyx.scipy.ndimage as ndimage
-
-
 import cupy as cp
 import numpy as np
 
 
-class TomoFunctions():
+class BackprojFunctions():
     def __init__(self, cl_conf):
 
         self.args = cl_conf.args
@@ -89,53 +86,6 @@ class TomoFunctions():
                 self.ne, self.ncproj, self.nz, self.args.dtype)  # note ncproj,nz!
         self.wfilter = self.cl_filter.calc_filter(self.args.fbp_filter)
 
-
-    def darkflat_correction(self, data, dark, flat):
-        """Dark-flat field correction"""
-
-        dark0 = dark.astype(self.args.dtype, copy=False)
-        flat0 = flat.astype(self.args.dtype, copy=False)
-        # works only for processing all angles
-        if self.args.flat_linear == 'True' and data.shape[0] == self.nproj:
-            flat0_p0 = cp.mean(flat0[:flat0.shape[0]//2], axis=0)
-            flat0_p1 = cp.mean(flat0[flat0.shape[0]//2+1:], axis=0)
-            v = cp.linspace(0, 1, self.nproj)[..., cp.newaxis, cp.newaxis]
-            flat0 = (1-v)*flat0_p0+v*flat0_p1
-        else:
-            flat0 = cp.mean(flat0, axis=0)
-        dark0 = cp.mean(dark0, axis=0)
-        res = (data.astype(self.args.dtype, copy=False)-dark0) / (flat0-dark0+1e-3)
-        res[res <= 0] = 1
-        return res
-
-    def minus_log(self, data):
-        """Taking negative logarithm"""
-
-        data[:] = -cp.log(data)
-        data[cp.isnan(data)] = 6.0
-        data[cp.isinf(data)] = 0
-        return data  # reuse input memory
-
-    def remove_outliers(self, data):
-        """Remove outliers"""
-
-        if (int(self.args.dezinger) > 0):
-            w = int(self.args.dezinger)
-            if len(data.shape) == 3:
-                fdata = ndimage.median_filter(data, [w, 1, w])
-            else:
-                fdata = ndimage.median_filter(data, [w, w])
-            data[:] = cp.where(cp.logical_and(
-                data > fdata, (data - fdata) > self.args.dezinger_threshold), fdata, data)
-        return data
-        # if(int(self.args.dezinger) > 0):
-        #     r = int(self.args.dezinger)
-        #     fdata = ndimage.median_filter(data, [1, r, r])
-        #     ids = cp.where(cp.abs(fdata-data) > 0.5*cp.abs(fdata))
-        #     data[ids] = fdata[ids]
-        # return data
-
-
     def fbp_filter_center(self, data, sht=0):
         """FBP filtering of projections with applying the rotation center shift wrt to the origin"""
 
@@ -151,70 +101,3 @@ class TomoFunctions():
         data[:] = tmp[:, :, self.ne//2-self.n//2:self.ne//2+self.n//2]
 
         return data  # reuse input memory
-
-    def pad360(self, data):
-        """Pad data with 0 to handle 360 degrees scan"""
-
-        if (self.centeri < self.ni//2):
-            # if rotation center is on the left side of the ROI
-            data[:] = data[:, :, ::-1]
-        w = max(1, int(2*(self.ni-self.center)))
-
-        if self.args.pad_endpoint == 'True':
-            v = cp.linspace(1, 0, w, endpoint=True)
-        else:
-            v = cp.linspace(1, 0, w, endpoint=False)
-        v = v**5*(126-420*v+540*v**2-315*v**3+70*v**4)
-        data[:, :, -w:] *= v
-
-        # double sinogram size with adding 0
-        data = cp.pad(data, ((0, 0), (0, 0), (0, data.shape[-1])), 'constant')
-        return data
-
-    def proc_sino(self, data, dark, flat, res=None):
-        """Processing a sinogram data chunk"""
-
-        if not isinstance(res, cp.ndarray):
-            res = cp.zeros(data.shape, self.args.dtype)
-        # dark flat field correrction
-        data[:] = self.remove_outliers(data)
-        dark[:] = self.remove_outliers(dark)
-        flat[:] = self.remove_outliers(flat)
-        res[:] = self.darkflat_correction(data, dark, flat)
-        # remove stripes
-        if self.args.remove_stripe_method == 'fw':
-            res[:] = remove_stripe.remove_stripe_fw(
-                res, self.args.fw_sigma, self.args.fw_filter, self.args.fw_level)
-        elif self.args.remove_stripe_method == 'ti':
-            res[:] = remove_stripe.remove_stripe_ti(
-                res, self.args.ti_beta, self.args.ti_mask)
-        elif self.args.remove_stripe_method == 'vo-all':
-            res[:] = remove_stripe.remove_all_stripe(
-                res, self.args.vo_all_snr, self.args.vo_all_la_size, self.args.vo_all_sm_size, self.args.vo_all_dim)
-
-        return res
-
-    def proc_proj(self, data, res=None):
-        """Processing a projection data chunk"""
-
-        if not isinstance(res, cp.ndarray):
-            res = cp.zeros(
-                [data.shape[0], data.shape[1], self.n], self.args.dtype)
-        # retrieve phase
-        if self.args.retrieve_phase_method == 'Gpaganin' or self.args.retrieve_phase_method == 'paganin':
-            data[:] = retrieve_phase.paganin_filter(
-                data,  self.args.pixel_size*1e-4, self.args.propagation_distance/10, self.args.energy, \
-                self.args.retrieve_phase_alpha, self.args.retrieve_phase_method, self.args.retrieve_phase_delta_beta, \
-                self.args.retrieve_phase_W*1e-4)
-        if self.args.rotate_proj_angle != 0:
-            data[:] = adjust_projections.rotate(
-                data, self.args.rotate_proj_angle, self.args.rotate_proj_order)
-        # minus log
-        if self.args.minus_log == 'True':
-            data[:] = self.minus_log(data)
-        # padding for 360 deg recon
-        if (self.args.file_type == 'double_fov'):
-            res[:] = self.pad360(data)
-        else:
-            res[:] = data[:]
-        return res
